feat:#20-Button CSR separation
