+++ src/app/page.tsx
@@ -1,7 +1,3 @@
-"use client";
-
-import { ButtonWrapper } from "@/components/common/headless/Button";
-import { useRouter } from "next/navigation";
 import Image from "next/image";
 import sm_1 from "@/../public/assets/landing/img_landing_sm_01.svg";
 import sm_2 from "@/../public/assets/landing/img_landing_sm_02.svg";
@@ -9,18 +5,9 @@ import sm_3 from "@/../public/assets/landing/img_landing_sm_03.svg";
 import lg_1 from "@/../public/assets/landing/img_landing_lg_01.svg";
 import lg_2 from "@/../public/assets/landing/img_landing_lg_02.svg";
 import lg_3 from "@/../public/assets/landing/img_landing_lg_03.svg";
+import LandingClient from "@/pages/LandingClient";
 
 export default function Home() {
-  const router = useRouter();
-
-  const handleSignInNavigation = () => {
-    router.push("/sign-in");
-  };
-
-  const handleSignUpNavigation = () => {
-    router.push("/sign-up");
-  };
-
   return (
     <div className="w-full min-h-screen flex items-center flex-col lg:py-[8rem] sm:py-[6.5rem] bg-background-400">
       <p className="text-black lg:text-[3.6rem] sm:text-[2.4rem] font-semibold lg:leading-[5rem] sm:leading-[3.4rem] text-center">
@@ -104,18 +91,7 @@ export default function Home() {
           </div>
         </div>
       </div>
-      <div className="flex lg:flex-row sm:flex-col lg:gap-[1.6rem] sm:gap-[0.8rem]">
-        <ButtonWrapper id="move-signin" onClick={handleSignInNavigation}>
-          <ButtonWrapper.Button className="lg:w-[34rem] lg:h-[6.4rem] sm:w-[32.7rem] sm:h-[5.4rem] rounded-[5rem] p-[1.6rem] font-semibold lg:text-[2rem] sm:text-[1.6rem] lg:leading-[3.2rem] sm:leading-[2.6rem] flex items-center justify-center text-white">
-            로그인
-          </ButtonWrapper.Button>
-        </ButtonWrapper>
-        <ButtonWrapper id="move-signin" onClick={handleSignUpNavigation}>
-          <ButtonWrapper.Button className="lg:w-[34rem] lg:h-[6.4rem] sm:w-[32.7rem] sm:h-[5.4rem] rounded-[5rem] p-[1.6rem] font-semibold lg:text-[2rem] sm:text-[1.6rem] lg:leading-[3.2rem] sm:leading-[2.6rem] flex items-center justify-center text-blue-300 bg-white border border-blue-300">
-            회원가입
-          </ButtonWrapper.Button>
-        </ButtonWrapper>
-      </div>
+      <LandingClient />
     </div>
   );
 }
